Rimossa applicazione segmenti_test e refactoring minore dei test
