--- segmenti/segmenti_test/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class SegmentiTestConfig(AppConfig):
-    name = 'segmenti_test'
